backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
joblib==1.3.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
scapy==2.5.0
requests==2.31.0
